+++ src/framework/behavior/__init__.lua
@@ -1,0 +1,3 @@
+local require = require
+require 'framework.behavior.load'
+return require 'framework.behavior.tree'